+++ src/generator/with_rand09.rs
@@ -1,4 +1,4 @@
-//! Integration with `rand` (v0.8) crate.
+//! Integration with `rand` (v0.9) crate.
 
 #![cfg(feature = "rand")]
 #![cfg_attr(docsrs, doc(cfg(feature = "rand")))]
@@ -18,17 +18,21 @@ impl<T: RngCore> Scru128Rng for Adapter<T> {
 
 impl<T: RngCore> Scru128Generator<Adapter<T>> {
     /// Creates a generator object with a specified random number generator that implements
-    /// [`RngCore`] from `rand` (v0.8) crate. The specified random number generator should be
+    /// [`RngCore`] from `rand` (v0.9) crate. The specified random number generator should be
     /// cryptographically strong and securely seeded.
-    pub const fn with_rand08(rng: T) -> Self {
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// # #[cfg(feature = "default_rng")]
+    /// # {
+    /// use scru128::Scru128Generator;
+    ///
+    /// let mut g = Scru128Generator::with_rand09(rand::rng());
+    /// println!("{}", g.generate());
+    /// # }
+    /// ```
+    pub const fn with_rand09(rng: T) -> Self {
         Self::with_rng(Adapter(rng))
-    }
-}
-
-/// This is a deprecated blanket impl retained for backward compatibility. Do not depend on this
-/// impl; use [`Scru128Generator::with_rand08()`] instead.
-impl<T: RngCore> Scru128Rng for T {
-    fn next_u32(&mut self) -> u32 {
-        self.next_u32()
     }
 }